fix config replace bug

+++ Makefile
@@ -30,7 +30,7 @@ install-core: install-polkit install-cli
 	install core/functions.sh $(DESTDIR)/usr/lib/sulin/dsl/
 	install core/debian.svg $(DESTDIR)/usr/lib/sulin/dsl/
 	install core/hostctl $(DESTDIR)/usr/lib/sulin/dsl/
-	install debian.conf  $(DESTDIR)/etc/
+	[ ! -f $(DESTDIR)/etc/debian.conf ] && install debian.conf  $(DESTDIR)/etc/
 
 install-terminal:
 	mkdir -p $(DESTDIR)/usr/share/applications/ || true
